==> 04-Data-Enrichment.ipynb <==
### Libraries

import numpy as np
import pandas as pd
import pickle

### Get Data

with open('./data/processed/carrier_aircraft_combo_capacity_dict.txt', 'rb') as handle:
    carrier_aircraft_combo_capacity = pickle.loads(handle.read()) 
dat = pd.read_pickle('./data/processed/final/all-sched.pkl')
dat.head()

#### Weekly Flights for Schedule
Adding a column, <i>weekly_flights</i>, to the dataframe that translates the frequency information into weekly number of flights.

dat['weekly_flights'] = dat.apply(lambda row: len(row.frequency), axis = 1)
dat.insert(4, 'weekly_flights', dat.pop('weekly_flights'))
dat.head()

#### Aircraft Capacity
Adding a column, <i>capacity</i>, to the dataframe that captures the capacity of the aircraft flying the schedule.

dat['craft_capacity'] = dat.apply(lambda row: 
                                     carrier_aircraft_combo_capacity[row.operator + '-' + row.aircraft], 
                                     axis = 1)
dat.insert(3, 'craft_capacity', dat.pop('craft_capacity'))

dat.head()

#### Add <i>weekly_capacity</i> and <i>weekly_est_pass</i>
For every schedule, adding columns:
<ul>
    <li>weekly_capacity = weekly_flights*capacity
    <li>weekly_est_pass = weekly_capacity*carrier_plf

dat['weekly_capacity'] = dat.apply(lambda row: row.weekly_flights*row.craft_capacity, axis = 1)
dat.insert(6, 'weekly_capacity', dat.pop('weekly_capacity'))

with open('./data/processed/carrier_plf_dict.txt', 'rb') as handle:
    carrier_plf = pickle.loads(handle.read())
dat['weekly_est_pass'] = dat.apply(lambda row: round(row.weekly_capacity*carrier_plf[row.operator], 0), 
                                   axis = 1)
dat.insert(7, 'weekly_est_pass', dat.pop('weekly_est_pass'))

dat.head()

#### Add <i>from_time_slot</i> and <i>to_time_slot</i>
Divide the day (0.0 hrs to 24.0 hrs) into 48 serially numbered time-slots of 0.5 hours each.
For each schedule, insert columns from_time_slot and to_time_slot that identify which serially numbered time slot the schedule's <i>from_time</i> and <i>to_time</i> belong to.

dat.insert(10, 'from_time_slot', 0)
dat.insert(13, 'to_time_slot', 0)

for i in range(dat.shape[0]):
    hr_, min_ = divmod(dat.from_time[i], 1)
    if min_ < 0.5: dat.loc[i, 'from_time_slot'] = 2*hr_ + 1
    else: dat.loc[i, 'from_time_slot'] = 2*hr_ + 2
        
    hr_, min_ = divmod(dat.to_time[i], 1)
    if min_ < 0.5: dat.loc[i, 'to_time_slot'] = 2*hr_ + 1
    else: dat.loc[i, 'to_time_slot'] = 2*hr_ + 2

dat.head()

#### One-Hot-Encode <i>Frequency</i>
Adding 7 new columns, one for each day of the week and if a particular schedule flies on a given day of the week, set value in the appropriate column to True.

for i in range(1, 8, 1):
    col_name = 'day_' + str(i)
    dat[col_name] = False
    
for i in range(dat.shape[0]):
    for j in range(1, 8, 1):
        col_name = 'day_' + str(j)
        if j in dat['frequency'][i]:
            dat.loc[i, col_name] = True
            
dat.head()

dat.to_pickle('./data/processed/final/all-sched-enriched.pkl')
dat.to_excel('./data/processed/final/all-sched-enriched.xlsx', index = False)
dat.to_csv('./data/processed/final/all-sched-enriched.csv', index = False)
